% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostat_data.R
\name{biostat_data}
\alias{biostat_data}
\title{Data from Methods in Biostatistics in R}
\usage{
biostat_data(...)
}
\arguments{
\item{...}{character files to retrieve.  See examples}
}
\value{
A character vector of the filename.
}
\description{
Data from Methods in Biostatistics in R
}
\examples{
biostat_data("bmi_age.txt")
biostat_data("FLAIR.nii.gz")
ff = system.file("extdata", "shhs1.txt", package = "biostatmethods")
if (file.exists(ff)) {
   file.remove(ff)
}
testthat::expect_error(biostat_data("shhs1.txt"))
token = nsrr::nsrr_token()
if (!is.null(token)) {
   if (biostat_download_shhs()) {
      biostat_data("shhs1.txt")
   }
}
}
