% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostat_data.R
\name{biostat_data}
\alias{biostat_data}
\title{Data from Methods in Biostatistics in R}
\usage{
biostat_data(...)
}
\arguments{
\item{...}{character files to retrieve.  See examples}
}
\value{
A character vector of the filename.
}
\description{
Data from Methods in Biostatistics in R
}
\examples{
biostat_data("bmi_age.txt")
biostat_data("FLAIR.nii.gz")
}
