% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostat_download_shhs.R
\name{biostat_download_shhs}
\alias{biostat_download_shhs}
\title{Download SHHS data}
\usage{
biostat_download_shhs(token = nsrr::nsrr_token(),
  out_dir = system.file(package = "biostatmethods"))
}
\arguments{
\item{token}{Token for 'NSRR' data resource, to pass to}

\item{out_dir}{Directory to download the file}
}
\value{
A logical if the data was downloaded
}
\description{
Download SHHS data
}
\examples{
biostat_download_shhs()
}
